+++ frontend/app/features/archive/components/ProductionInfoSection.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useEffect, useState } from "react";
-import ComplexEditableField from "~/shared/components/ArchiveRichTextFieldWrapper";
+import ComplexEditableField from "~/shared/components/ComplexEditableField";
 
 type ProductionInfoSectionProps = {
   tagline: string;
